refactor(monitor): extract live network views into internal/services/monitor
Dashboard / Connections / DeviceActivity + device traces + tcpdump captures
(app/{monitor,trace,pcap}.go) move into internal/services/monitor.Service, which
takes *config.Config, *store.Store and *proxy.Service (the dashboard surfaces the
two Telegram proxies' status). App holds *monitor.Service + thin delegators (+ the
DashboardView/Trace/Pcap type aliases).

The tcpdump plumbing shared by pcap capture and ClientHello blob capture
(tcpdumpPath / deviceIface / ErrNeedTcpdump / opkg install) moves to a new
internal/tools/tcpdump package so neither service depends on the other; app keeps
ErrNeedTcpdump + InstallPackage as a re-export + delegator so the server's
errors.Is(err, app.ErrNeedTcpdump) is unchanged. blobcap.go (still in app, moves
with the blobs service next) now uses the tcpdump package directly.

internal/server handlers + JSON unchanged. Verified: go build (linux) + go vet +
go test green.
