+++ internal/app/monitor.go
@@ -1,104 +1,45 @@
 package app
 
-import "nfqws2strategy/internal/tools/netmon"
+import (
+	"nfqws2strategy/internal/services/monitor"
+	"nfqws2strategy/internal/tools/tcpdump"
+)
 
-// DashboardView is the home-page snapshot: TG WS proxy status, conntrack totals,
-// the DPI engine's NFQUEUE stats, and WAN byte counters (the UI derives a rate
-// by diffing successive samples).
-type DashboardView struct {
-	TGWS   TGWSStatus   `json:"tgws"`
-	Socks5 Socks5Status `json:"socks5"`
+// Live network monitoring (dashboard, connections, device activity, device
+// traces, tcpdump captures) now lives in internal/services/monitor. These type
+// aliases + thin delegators keep the App surface the HTTP handlers call
+// unchanged.
 
-	// Nfqws2Running is true when the DPI engine's NFQUEUE (MainQueue) is bound.
-	Nfqws2Running bool `json:"nfqws2_running"`
+type (
+	DashboardView      = monitor.DashboardView
+	ConnectionsView    = monitor.ConnectionsView
+	DeviceActivityView = monitor.DeviceActivityView
+	Trace              = monitor.Trace
+	TraceEvent         = monitor.TraceEvent
+	TraceConn          = monitor.TraceConn
+	Pcap               = monitor.Pcap
+)
 
-	Conntrack struct {
-		Count int `json:"count"`
-		Max   int `json:"max"`
-	} `json:"conntrack"`
+// ErrNeedTcpdump re-exports the shared sentinel so the server's
+// errors.Is(err, app.ErrNeedTcpdump) keeps working for both the device pcap
+// capture and the ClientHello blob capture.
+var ErrNeedTcpdump = tcpdump.ErrNeedInstall
 
-	Conns struct {
-		Total   int            `json:"total"`
-		Failing int            `json:"failing"`
-		ByProto map[string]int `json:"by_proto"`
-	} `json:"conns"`
+func (a *App) Dashboard(host string) DashboardView         { return a.monitor.Dashboard(host) }
+func (a *App) Connections() (ConnectionsView, error)       { return a.monitor.Connections() }
+func (a *App) DeviceActivity() (DeviceActivityView, error) { return a.monitor.DeviceActivity() }
 
-	Queues    []netmon.QueueStat  `json:"queues"`
-	MainQueue int                 `json:"main_queue"`
-	WAN       []netmon.IfaceBytes `json:"wan"`
+func (a *App) StartDeviceTrace(ip string, seconds int) (*Trace, error) {
+	return a.monitor.StartDeviceTrace(ip, seconds)
 }
+func (a *App) GetTrace(id string) (*Trace, bool) { return a.monitor.GetTrace(id) }
 
-// Dashboard assembles the home view. The TG WS card works on any platform; the
-// /proc-derived sections are best-effort and simply stay empty when unavailable
-// (e.g. on the non-Linux dev box), so the page always renders.
-func (a *App) Dashboard(host string) DashboardView {
-	var d DashboardView
-	d.TGWS = a.TGWSStatusFor(host)
-	d.Socks5 = a.Socks5StatusFor(host)
-	d.MainQueue = a.Cfg.MainQueue
-	d.Conns.ByProto = map[string]int{}
-
-	if cur, limit, err := netmon.Count(); err == nil {
-		d.Conntrack.Count = cur
-		d.Conntrack.Max = limit
-	}
-	if conns, err := netmon.Conntrack(); err == nil {
-		d.Conns.Total = len(conns)
-		for _, c := range conns {
-			d.Conns.ByProto[c.Proto]++
-			if c.Failing() {
-				d.Conns.Failing++
-			}
-		}
-	}
-	if qs, err := netmon.Queues(); err == nil {
-		d.Queues = qs
-		for _, q := range qs {
-			if q.Queue == a.Cfg.MainQueue {
-				d.Nfqws2Running = true
-				break
-			}
-		}
-	}
-	if ifs, err := netmon.Ifaces(); err == nil {
-		want := make(map[string]bool, len(a.Cfg.WANIfaces))
-		for _, n := range a.Cfg.WANIfaces {
-			want[n] = true
-		}
-		for _, f := range ifs {
-			if want[f.Iface] {
-				d.WAN = append(d.WAN, f)
-			}
-		}
-	}
-	return d
+func (a *App) StartPcap(ip string, seconds int) (*Pcap, error) {
+	return a.monitor.StartPcap(ip, seconds)
 }
+func (a *App) GetPcap(id string) (*Pcap, bool)                 { return a.monitor.GetPcap(id) }
+func (a *App) PcapFile(id string) (path, name string, ok bool) { return a.monitor.PcapFile(id) }
 
-// ConnectionsView is the full live connection list for the Connections tab
-// (filtered/sorted/paginated client-side — the table is small).
-type ConnectionsView struct {
-	Items []netmon.Conn `json:"items"`
-	Count int           `json:"count"`
-}
-
-func (a *App) Connections() (ConnectionsView, error) {
-	conns, err := netmon.Conntrack()
-	if err != nil {
-		return ConnectionsView{}, err
-	}
-	return ConnectionsView{Items: conns, Count: len(conns)}, nil
-}
-
-// DeviceActivityView groups live connections by LAN device.
-type DeviceActivityView struct {
-	Devices []netmon.Device `json:"devices"`
-}
-
-func (a *App) DeviceActivity() (DeviceActivityView, error) {
-	conns, err := netmon.Conntrack()
-	if err != nil {
-		return DeviceActivityView{}, err
-	}
-	arp, _ := netmon.ARP() // best-effort: enriches MAC/bridge, not required
-	return DeviceActivityView{Devices: netmon.GroupDevices(conns, arp)}, nil
-}
+// InstallPackage installs a whitelisted opkg package (currently only tcpdump),
+// surfaced by the device-capture install prompt.
+func (a *App) InstallPackage(pkg string) (string, error) { return tcpdump.Install(pkg) }
